backend/requirements.txt
fastapi==0.127.1
uvicorn[standard]==0.32.0
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
alembic==1.14.0
celery==5.4.0
redis==5.2.0
pdfplumber==0.11.4
pydantic==2.10.0
pydantic-settings==2.6.1
python-multipart==0.0.17
httpx==0.28.0
pytest==8.3.3
beautifulsoup4==4.12.3
google-genai==0.8.0